move start func
